Добавил контроллер и абстрактный класс
