+++ src/components/AbstractComponent.js
@@ -1,0 +1,21 @@
+import {createElement} from '../components/utils';
+
+class AbstractComponent {
+  constructor() {
+    this._element = null;
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  getTemplate() {
+    throw Error(`Abstract method not implemented`);
+  }
+}
+
+export {AbstractComponent};
