+++ src/components/TaskList.js
@@ -1,0 +1,7 @@
+import {AbstractComponent} from './AbstractComponent';
+class TaskList extends AbstractComponent {
+  getTemplate() {
+    return `<div class="board__tasks"></div>`;
+  }
+}
+export {TaskList};
