--- src/components/boardFilter.js
@@ -1,9 +1,0 @@
-const boardFilter = () => {
-  return `<div class="board__filter-list">
-  <a href="#" class="board__filter">SORT BY DEFAULT</a>
-  <a href="#" class="board__filter">SORT BY DATE up</a>
-  <a href="#" class="board__filter">SORT BY DATE down</a>
-  </div>`.trim();
-};
-
-export {boardFilter};
