+++ src/components/task.js
@@ -1,6 +1,6 @@
-import {createElement} from './utils';
+import {AbstractComponent} from './AbstractComponent';
 
-class Task {
+class Task extends AbstractComponent {
   constructor({
     description,
     dueDate,
@@ -8,22 +8,13 @@ class Task {
     color,
     repeatingDays
   }) {
+    super();
     this._description = description;
     this._dueDate = new Date(dueDate);
     this._tags = tags;
     this._color = color;
-    this._element = null;
     this._repeatingDays = repeatingDays;
   }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
   getTemplate() {
     return `<article class="card card--${this._color} ${Object.values(this._repeatingDays).some((it) => it === true) ? `card--repeat` : `` }">
               <div class="card__form">
@@ -83,4 +74,4 @@ class Task {
   }
 }
 
-export default Task;
+export {Task};
