+++ src/main.js
@@ -1,26 +1,18 @@
-import Task from './components/task';
+import {BoardController} from './controllers/BoardController';
+import {getTask} from './components/data';
 import {menu} from './components/menu';
 import {search} from './components/search';
 import {mainFilter} from './components/mainFilter';
-import {boardFilter} from './components/boardFilter';
-import {board} from './components/board';
 import {loadMore} from './components/loadMore';
-import TaskEdit from './components/editTask';
-import {getTask} from './components/data';
-import {Position, render, createElement} from './components/utils';
-
+import {Position, createElement, render} from './components/utils';
 
 const TASK_COUNT = 3;
-
-
 const taskMocks = new Array(TASK_COUNT)
   .fill(``)
   .map(getTask);
 
-
 const menuContainer = document.querySelector(`.main__control`);
 const siteMainContainer = document.querySelector(`.main`);
-
 
 const renderMenu = () => {
   render(menuContainer, createElement(menu()), Position.BEFOREEND);
@@ -31,63 +23,17 @@ const renderSearch = () => {
 const renderMainFilter = () => {
   render(siteMainContainer, createElement(mainFilter()), Position.BEFOREEND);
 };
-const renderBoard = () => {
-  render(siteMainContainer, createElement(board()), Position.BEFOREEND);
-};
-
 
 renderMenu();
 renderSearch();
 renderMainFilter();
-renderBoard();
 
 const boardContainer = siteMainContainer.querySelector(`.board`);
 const tasksContainer = siteMainContainer.querySelector(`.board__tasks`);
-const renderBoardFilter = () => {
-  render(boardContainer, createElement(boardFilter()), Position.AFTERBEGIN);
-};
-
-const renderTask = (taskMock) => {
-  const task = new Task(taskMock);
-  const taskEdit = new TaskEdit(taskMock);
-
-  const onEscKeyDown = (evt) => {
-    if (evt.key === `Escape` || evt.key === `Esc`) {
-      tasksContainer.replaceChild(task.getElement(), taskEdit.getElement());
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    }
-  };
-
-  task.getElement()
-    .querySelector(`.card__btn--edit`)
-    .addEventListener(`click`, () => {
-      tasksContainer.replaceChild(taskEdit.getElement(), task.getElement());
-      document.addEventListener(`keydown`, onEscKeyDown);
-    });
-
-  taskEdit.getElement().querySelector(`textarea`)
-    .addEventListener(`focus`, () => {
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    });
-
-  taskEdit.getElement().querySelector(`textarea`)
-    .addEventListener(`blur`, () => {
-      document.addEventListener(`keydown`, onEscKeyDown);
-    });
-
-  taskEdit.getElement()
-    .querySelector(`.card__save`)
-    .addEventListener(`click`, () => {
-      tasksContainer.replaceChild(task.getElement(), taskEdit.getElement());
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    });
-
-  render(tasksContainer, task.getElement(), Position.BEFOREEND);
-};
+const boardController = new BoardController(tasksContainer, taskMocks);
+boardController.init();
 const renderLoadMore = () => {
   render(boardContainer, createElement(loadMore()), Position.BEFOREEND);
 };
-
-renderBoardFilter();
-taskMocks.forEach((taskMock) => renderTask(taskMock));
 renderLoadMore();
+
